auth_bot/requirements.txt
pyrogram==2.0.106
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-jose==3.3.0
razorpay==1.4.1
paypalrestsdk==1.13.3
cryptography==41.0.8
aioredis==2.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
